MainPackage/requirements.txt
allure_python_commons==2.9.45
openpyxl==3.0.9
pytest==7.1.1
selenium==4.2.0
webdriver_manager==3.5.2
